Split command tests out to indavidual test files
